refactor: standardize partner-performance and partnership-program-implementation layouts
Partner Performance (340 → 111 lines, 67% reduction):
- Create OverviewTab, TiersTab, TopPerformersTab components
- Create StrategicInsights component
- Refactor page.tsx to use wrapper components
- Follows Pattern B (Tabbed) standard

Partnership Program Implementation (308 → 48 lines, 84% reduction):
- Create AchievementsGrid component
- Create TechnicalDetails component
- Create NarrativeSections component
- Refactor page.tsx to use wrapper components
- Follows Pattern A (Simple) standard

Both projects now follow standardized patterns for improved maintainability.
