+++ src/app/projects/partner-performance/page.tsx
@@ -1,93 +1,28 @@
 'use client'
 export const dynamic = 'force-static'
 
-import dynamicImport from 'next/dynamic'
 import { TrendingUp, Users, Target, DollarSign } from 'lucide-react'
 import { useQueryState } from 'nuqs'
 
 import { ProjectPageLayout } from '@/components/projects/project-page-layout'
 import { MetricsGrid } from '@/components/projects/metrics-grid'
 import { SectionCard } from '@/components/ui/section-card'
-import { ChartContainer } from '@/components/ui/chart-container'
 import { formatCurrency, formatNumber, formatPercentage } from '@/lib/utils/data-formatters'
 import { NarrativeSections } from './components/NarrativeSections'
-
-// Lazy-load chart components with Suspense fallback
-const PartnerTierChart = dynamicImport(() => import('./PartnerTierChart'), {
-  loading: () => <div className="h-[var(--chart-height-md)] w-full animate-pulse bg-muted rounded-lg" />,
-  ssr: true,
-})
-const RevenueContributionChart = dynamicImport(() => import('./RevenueContributionChart'), {
-  loading: () => <div className="h-[var(--chart-height-md)] w-full animate-pulse bg-muted rounded-lg" />,
-  ssr: true,
-})
+import { OverviewTab } from './components/OverviewTab'
+import { TiersTab } from './components/TiersTab'
+import { TopPerformersTab } from './components/TopPerformersTab'
+import { StrategicInsights } from './components/StrategicInsights'
 
 // Real data based on CSV analysis: 83.2% win rate, $1.1M partner revenue
 const partnerMetrics = {
   totalPartners: 47,
   activePartners: 34,
-  certifiedPartners: 12,
-  totalRevenue: 1087420,
   partnerRevenue: 904387,
   partnerContribution: 83.2,
-  avgDealSize: 305,
   winRate: 83.2,
   quickRatio: 4.7, // SaaS benchmark >4
-  topPartnerRevenue: 287650, // 80/20 rule demonstration
 }
-
-const partnerTierPerformance = [
-  {
-    tier: 'Certified Partners',
-    count: 12,
-    revenue: 687420,
-    avgDeal: 485,
-    winRate: 89.4,
-    cycleLength: 73,
-    roi: 8.2,
-  },
-  {
-    tier: 'Legacy Partners',
-    count: 15,
-    revenue: 216967,
-    avgDeal: 298,
-    winRate: 78.1,
-    cycleLength: 89,
-    roi: 3.1,
-  },
-  {
-    tier: 'Inactive Partners',
-    count: 7,
-    revenue: 0,
-    avgDeal: 0,
-    winRate: 0,
-    cycleLength: 126,
-    roi: 0,
-  },
-  {
-    tier: 'New Partners',
-    count: 13,
-    revenue: 183033,
-    avgDeal: 267,
-    winRate: 71.2,
-    cycleLength: 105,
-    roi: 2.4,
-  },
-]
-
-const top80PercentPartners = [
-  { partner: 'TechFlow Solutions', revenue: 287650, deals: 89, winRate: 94.2, tier: 'Certified' },
-  {
-    partner: 'Digital Marketing Pro',
-    revenue: 184320,
-    deals: 67,
-    winRate: 87.6,
-    tier: 'Certified',
-  },
-  { partner: 'Revenue Boost Inc', revenue: 156780, deals: 54, winRate: 91.1, tier: 'Certified' },
-  { partner: 'Growth Partners LLC', revenue: 142890, deals: 48, winRate: 85.4, tier: 'Legacy' },
-  { partner: 'Channel Dynamics', revenue: 133247, deals: 45, winRate: 82.2, tier: 'Legacy' },
-]
 
 const tabs = ['overview', 'tiers', 'top-performers'] as const
 type Tab = (typeof tabs)[number]
@@ -152,189 +87,25 @@ export default function PartnerPerformanceIntelligence() {
         setActiveTab(tab)
       }}
     >
-          {/* Key Metrics using standardized MetricsGrid */}
-          <MetricsGrid metrics={metrics} columns={4} className="mb-8" />
+      {/* Key Metrics using standardized MetricsGrid */}
+      <MetricsGrid metrics={metrics} columns={4} className="mb-8" />
 
-          {/* Tab Content wrapped in SectionCard */}
-          <SectionCard
-            title="Partner Performance Analysis"
-            description="Comprehensive partner analytics and performance insights"
-            className="mb-8"
-          >
-            {activeTab === 'overview' && (
-              <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                <ChartContainer
-                  title="Partner Tier Performance Analysis"
-                  description="Revenue and ROI analysis across certified, legacy, and new partner tiers"
-                  height={250}
-                >
-                  <PartnerTierChart />
-                </ChartContainer>
+      {/* Tab Content wrapped in SectionCard */}
+      <SectionCard
+        title="Partner Performance Analysis"
+        description="Comprehensive partner analytics and performance insights"
+        className="mb-8"
+      >
+        {activeTab === 'overview' && <OverviewTab />}
+        {activeTab === 'tiers' && <TiersTab />}
+        {activeTab === 'top-performers' && <TopPerformersTab />}
+      </SectionCard>
 
-                <ChartContainer
-                  title="Channel Revenue Contribution"
-                  description="Partner revenue distribution following industry 80/20 performance rule"
-                  height={250}
-                >
-                  <RevenueContributionChart />
-                </ChartContainer>
-              </div>
-            )}
+      {/* Strategic Insights */}
+      <StrategicInsights winRate={partnerMetrics.winRate} quickRatio={partnerMetrics.quickRatio} />
 
-            {activeTab === 'tiers' && (
-              <div>
-                <h3 className="text-lg font-semibold mb-4">
-                  Partner Tier Intelligence & ROI Metrics
-                </h3>
-                <p className="text-sm text-muted-foreground mb-6">
-                  Comprehensive partner performance analysis across certification levels and
-                  engagement tiers
-                </p>
-
-                <div className="overflow-x-auto">
-                  <table className="w-full">
-                    <thead>
-                      <tr className="border-b border-border">
-                        <th className="text-left py-3 px-4 font-semibold">Partner Tier</th>
-                        <th className="text-left py-3 px-4 font-semibold">Count</th>
-                        <th className="text-left py-3 px-4 font-semibold">Revenue</th>
-                        <th className="text-left py-3 px-4 font-semibold">Avg Deal</th>
-                        <th className="text-left py-3 px-4 font-semibold">Win Rate</th>
-                        <th className="text-left py-3 px-4 font-semibold">Cycle (Days)</th>
-                        <th className="text-left py-3 px-4 font-semibold">ROI</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {partnerTierPerformance.map((tier, index) => (
-                        <tr
-                          key={index}
-                          className="border-b border-border/50 hover:bg-muted/50 transition-colors"
-                        >
-                          <td className="py-4 px-4 font-medium">{tier.tier}</td>
-                          <td className="py-4 px-4">{formatNumber(tier.count)}</td>
-                          <td className="py-4 px-4">
-                            {formatCurrency(tier.revenue, { compact: true })}
-                          </td>
-                          <td className="py-4 px-4">{formatCurrency(tier.avgDeal)}</td>
-                          <td className="py-4 px-4">
-                            <span
-                              className={`px-2 py-1 rounded-full text-xs font-medium ${
-                                tier.winRate >= 85
-                                  ? 'bg-success/20 text-success'
-                                  : tier.winRate >= 70
-                                    ? 'bg-warning/20 text-warning'
-                                    : 'bg-destructive/20 text-destructive'
-                              }`}
-                            >
-                              {formatPercentage(tier.winRate / 100)}
-                            </span>
-                          </td>
-                          <td className="py-4 px-4">{formatNumber(tier.cycleLength)}</td>
-                          <td className="py-4 px-4">
-                            <span
-                              className={`px-2 py-1 rounded-full text-xs font-medium ${
-                                tier.roi >= 5
-                                  ? 'bg-success/20 text-success'
-                                  : tier.roi >= 3
-                                    ? 'bg-warning/20 text-warning'
-                                    : 'bg-destructive/20 text-destructive'
-                              }`}
-                            >
-                              {tier.roi}x
-                            </span>
-                          </td>
-                        </tr>
-                      ))}
-                    </tbody>
-                  </table>
-                </div>
-              </div>
-            )}
-
-            {activeTab === 'top-performers' && (
-              <div>
-                <h3 className="text-lg font-semibold mb-4">
-                  Top 80% Revenue Partners (Pareto Principle)
-                </h3>
-                <p className="text-sm text-muted-foreground mb-6">
-                  Elite partner performance driving majority of channel revenue following the proven
-                  80/20 distribution rule
-                </p>
-
-                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-                  {top80PercentPartners.map((partner, index) => (
-                    <div key={index} className="bg-card border border-border rounded-xl p-4">
-                      <div className="flex items-center justify-between mb-3">
-                        <h4 className="text-base font-semibold">{partner.partner}</h4>
-                        <span
-                          className={`px-2 py-1 rounded-full text-xs font-medium ${
-                            partner.tier === 'Certified'
-                              ? 'bg-primary/20 text-primary'
-                              : 'bg-secondary/20 text-secondary'
-                          }`}
-                        >
-                          {partner.tier}
-                        </span>
-                      </div>
-                      <div className="space-y-2">
-                        <div className="flex justify-between">
-                          <span className="text-muted-foreground">Revenue:</span>
-                          <span className="font-medium">
-                            {formatCurrency(partner.revenue, { compact: true })}
-                          </span>
-                        </div>
-                        <div className="flex justify-between">
-                          <span className="text-muted-foreground">Deals:</span>
-                          <span className="font-medium">{formatNumber(partner.deals)}</span>
-                        </div>
-                        <div className="flex justify-between">
-                          <span className="text-muted-foreground">Win Rate:</span>
-                          <span className="font-medium">
-                            {formatPercentage(partner.winRate / 100)}
-                          </span>
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )}
-          </SectionCard>
-
-          {/* Strategic Insights wrapped in SectionCard */}
-          <SectionCard
-            title="Partner Intelligence & Strategic Impact"
-            description="Key performance indicators and strategic insights"
-            className="mb-8"
-          >
-            <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-              <div className="bg-card border border-border rounded-xl p-4 text-center">
-                <div className="text-2xl font-semibold text-primary mb-1">
-                  {formatPercentage(partnerMetrics.winRate / 100)}
-                </div>
-                <div className="text-sm text-muted-foreground">
-                  Partner Channel Win Rate (Industry: 65-75%)
-                </div>
-              </div>
-              <div className="bg-card border border-border rounded-xl p-4 text-center">
-                <div className="text-2xl font-semibold text-secondary mb-1">
-                  {partnerMetrics.quickRatio}x
-                </div>
-                <div className="text-sm text-muted-foreground">
-                  SaaS Quick Ratio (Benchmark: &gt;4.0x)
-                </div>
-              </div>
-              <div className="bg-card border border-border rounded-xl p-4 text-center">
-                <div className="text-2xl font-semibold text-primary mb-1">80/20</div>
-                <div className="text-sm text-muted-foreground">
-                  Partner Revenue Distribution (Pareto Optimized)
-                </div>
-              </div>
-            </div>
-          </SectionCard>
-
-          {/* Professional Narrative Sections - STAR Method */}
-          <NarrativeSections />
+      {/* Professional Narrative Sections - STAR Method */}
+      <NarrativeSections />
     </ProjectPageLayout>
   )
 }
